+++ sample-apps/react-native/dogfood/src/constants/KnownUsers.ts
@@ -1,0 +1,27 @@
+export const KnownUsers = [
+  {
+    id: 'steve',
+    name: 'Steve Galilli',
+    image: 'https://ca.slack-edge.com/T02RM6X6B-U039J798FJ7-894cf1f07326-512',
+  },
+  {
+    id: 'khushal',
+    name: 'Khushal Agarwal',
+    image: 'https://ca.slack-edge.com/T02RM6X6B-U02DTREQ2KX-5d600c87d3bc-512',
+  },
+  {
+    id: 'santhosh',
+    name: 'Santhosh Vaiyapuri',
+    image: 'https://ca.slack-edge.com/T02RM6X6B-U0359AX2TUY-dc7dbec0bb88-512',
+  },
+  {
+    id: 'oliver',
+    name: 'Oliver Lazoroski',
+    image: 'https://ca.slack-edge.com/T02RM6X6B-U03HJKTMSQZ-cdf636547793-512',
+  },
+  {
+    id: 'zita',
+    name: 'Zita Szupera',
+    image: 'https://ca.slack-edge.com/T02RM6X6B-U02CA8MV9D1-8631020b96bf-512',
+  },
+];
